Do not execute query when nothing to delete
git-svn-id: https://bitoj.googlecode.com/svn/trunk/moodle/mod_programming@248 84fa3890-7848-0410-87fb-3d64402ee7d4

+++ lib.php
@@ -550,11 +550,13 @@ function programming_delete_old_submits($programmingid = -1, $userid = -1) {
         }
         $ids = implode(',', $ids);
     
-        #delete from submits table
-        execute_sql("DELETE FROM {$CFG->prefix}programming_submits WHERE id IN ({$ids})", false);
+        if ($ids) {
+            #delete from submits table
+            execute_sql("DELETE FROM {$CFG->prefix}programming_submits WHERE id IN ({$ids})", false);
 
-        #delete from test results table
-        execute_sql("DELETE FROM {$CFG->prefix}programming_test_results WHERE submitid IN ({$ids})", false);
+            #delete from test results table
+            execute_sql("DELETE FROM {$CFG->prefix}programming_test_results WHERE submitid IN ({$ids})", false);
+        }
     }
 
     return true;
